Some changes to bring population in line
